email_bot/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-bot"
version = "0.1.0"
description = "CLI tool to send personalized emails via Gmail API"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
    "Jinja2>=3.1.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
]

[project.scripts]
emailbot = "email_bot.cli:app_entry"

[tool.setuptools.packages.find]
where = ["."]
